# src/utils/api-spec_int.yaml
openapi: 3.0.0
info:
  title: WTM API 문서
  description: ''
  version: 1.0.0
paths:
...
  /my/{userId}:
    get:
      tags:
        - /my
      description: 유저 전용 메인 페이지를 불러옵니다
      operationId: getMypage
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: 조회 대상유저
      security:
        - bearerAuth: []
      responses:
        '200':
          description: '[완료] 유저 전용 페이지 로드를 완료하였습니다'
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  email:
                    type: string
                  name:
                    type: string
                  role:
                    type: string
                  address:
                    type: string
                  phone:
                    type: string
                  profilePicture:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
        '500':
          description: '[에러] 유저 데이터 전송에 문제가 생겨 페이지 로드에 실패하였습니다'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
  /my/settings/{userId}:
    get:
      tags:
        - /my
      description: 유저 정보를 불러옵니다
      operationId: getMySettings
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: 조회 대상유저
      security:
        - bearerAuth: []
      responses:
        '200':
          description: '[완료] 유저 정보를 로드하였습니다'
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  email:
                    type: string
                  name:
                    type: string
                  phone:
                    type: string
                  profilePicture:
                    type: string
                  role:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
        '500':
          description: '[에러] 유저 정보 요청에 문제가 생겨 데이터 로드에 실패하였습니다'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
    put:
      tags:
        - /my
      description: 유저 정보를 변경합니다
      operationId: updateUserSettings
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: 조회 대상유저
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                _id:
                  type: string
                email:
                  type: string
                password:
                  type: string
                name:
                  type: string
                phone:
                  type: string
                profilePicture:
                  type: string
                role:
                  type: string
      responses:
        '201':
          description: '[완료] 유저 정보 수정을 완료하였습니다'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
        '500':
          description: '[에러] 유저 정보 데이터 수정에 오류가 생겨 실패하였습니다'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
  /my/reviews/{userId}:
    get:
      tags:
        - /my
      description: 사용자가 작성한 리뷰 조회
      operationId: getUserReviews
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: 조회 대상유저
      security:
        - bearerAuth: []
      responses:
        '200':
          description: '[완료] 작성한 리뷰 정보를 불러왔습니다.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  storeId:
                    type: string
                  content:
                    type: string
                  revisit:
                    type: boolean
                  createdAt:
                    type: string
                    format: date-time
        '500':
          description: '[에러] 개인 리뷰 데이터를 불러오는 중 실패하였습니다.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
    delete:
      tags:
        - /my
      description: 사용자 리뷰 삭제하기
      operationId: deleteReview
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: 삭제할 리뷰 ID
      security:
        - bearerAuth: []
      responses:
        '204':
          description: '[완료] 해당 리뷰가 정상적으로 삭제되었습니다'
        '500':
          description: '[에러] 리뷰 삭제 중 오류가 발생하였습니다.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
...
  /my/tickets/{userId}:
    get:
      tags:
        - /my
      description: 사용자가 소유한 식권의 가게 조회
      operationId: getTicketsOwnedByUser
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: 조회 대상유저
      security:
        - bearerAuth: []
      responses:
        '200':
          description: '[완료] 사용자가 충전 사용한 티켓 정보를 불러왔습니다.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  ticketId:
                    type: string
                  amount:
                    type: integer
                  createdAt:
                    type: string
                    format: date-time
        '500':
          description: '[에러] 티켓 충전 사용 데이터를 불러오는 중 실패하였습니다.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
  /my/tickets/stores/charge:
    post:
      tags:
        - /my
      description: 사용자 티켓 충전
      operationId: purchaseMyTicket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                ticketId:
                  type: string
                amount:
                  type: integer
      responses:
        '201':
          description: '[완료] 사용자의 티켓 충전을 완료하였습니다.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
        '500':
          description: '[에러] 티켓 정보를 저장하던 중 실패하였습니다.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
  /my/tickets/{userId}/{storeId}/history:
    get:
      tags:
        - /my
      description: 사용자가 특정 가게의 티켓을 충전 사용한 내역 전체 조회
      operationId: getMyTicketHistoryByStore
      parameters:
        - in: path
          name: storeId
          required: true
          schema:
            type: string
          description: 특정 가게 지정
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: 사용자 지정
        - in: query
          name: month
          required: true
          schema:
            type: integer
          description: 조회할 달 (1~12)
        - in: query
          name: year
          required: true
          schema:
            type: integer
          description: 조회할 연도
      security:
        - bearerAuth: []
      responses:
        '200':
          description: '[완료] 사용자가 충전 사용한 티켓 정보를 불러왔습니다.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  storeId:
                    type: string
                  ticketId:
                    type: string
                  amount:
                    type: integer
                  createdAt:
                    type: string
                    format: date-time
        '500':
          description: '[에러] 티켓 충전 사용 데이터를 불러오는 중 실패하였습니다.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - jwt: []